moved new classes to sparsification package, rebased to latest base-modifier
