+++ src/sparseml/optim/__init__.py
@@ -22,7 +22,5 @@ Handles things like model pruning and increasing activation sparsity.
 from .analyzer import *
 from .helpers import *
 from .manager import *
-from .model_info import *
 from .modifier import *
-from .recipe_builder import *
 from .sensitivity import *
