+++ src/sparseml/sparsification/__init__.py
@@ -19,8 +19,12 @@ algorithms to models within in ML frameworks.
 
 # flake8: noqa
 
+from .analyzer import *
 from .info import *
+from .model_info import *
 from .modifier_epoch import *
 from .modifier_lr import *
 from .modifier_params import *
 from .modifier_pruning import *
+from .recipe_builder import *
+from .recipe_editor import *
